fix: rng and mintItem

+++ packages/contracts/src/libraries/GoldSpawnLogic.sol
@@ -43,9 +43,9 @@ library GoldSpawnLogic {
     pure
     returns (int128 x, int128 y)
   {
-    int256 rng = int256(random(seed));
-    x = int128(rng % 1000 - 500);
-    y = int128((rng / 1000) % 1000 - 500);
+    uint256 rng = random(seed);
+    x = int128(int256(rng % 1000) - 500);
+    y = int128(int256((rng / 1000) % 1000) - 500);
   }
 
   function random(uint256 seed) internal pure returns (uint256) {
